# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.\n\n## Rate Limiting\nLa API implementa limitación de solicitudes para garantizar disponibilidad:\n- **Endpoints públicos:** 60 solicitudes por minuto\n- **Endpoints administrativos:** 30 solicitudes por minuto\n\n## Caché\nAlgunos endpoints utilizan caché para mejorar el rendimiento:\n- **GET /personas:** caché de 5 minutos\n- **GET /empresas:** caché de 5 minutos\n- **GET /admin/estadisticas:** caché de 10 minutos\n\n## Tiempos de respuesta esperados\n- Operaciones de lectura: < 200ms\n- Operaciones de escritura: < 500ms",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas (CV ciego)",
                "description": "Obtiene talentos activos en formato de CV ciego (sin datos personales identificables).",
                "operationId": "getPersonas",
                "parameters": [
                    {
                        "name": "validado",
                        "in": "query",
                        "description": "Filtrar por validación",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "nivel_educacional",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "basica",
                                "media",
                                "tecnica",
                                "universitaria",
                                "postgrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PersonaCVCiego"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar nueva persona/talento",
                "description": "Crea un perfil de talento. El código se genera automáticamente.",
                "operationId": "createPersona",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona por ID",
                "operationId": "getPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Desactivar persona",
                "description": "Desactiva el perfil sin eliminarlo de la base de datos.",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona desactivada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/personas/{id}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Validar persona (solo administración)",
                "description": "Marca a una persona como validada para que aparezca en la vitrina.",
                "operationId": "validarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona validada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PersonaCVCiego": {
                "title": "Persona CV Ciego",
                "description": "Vista pública del talento sin datos personales identificables",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "PROV-2026-A1B2"
                    },
                    "resumen": {
                        "type": "string",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 5
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "PersonaInput": {
                "title": "Persona Input",
                "description": "Datos para crear o actualizar una persona",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    },
                    "resumen": {
                        "type": "string",
                        "example": "Contadora con experiencia en pymes y sector público",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "enum": [
                            "basica",
                            "media",
                            "tecnica",
                            "universitaria",
                            "postgrado"
                        ],
                        "example": "universitaria",
                        "nullable": true
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "example": "Contador Auditor",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "example": 2018,
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 7,
                        "nullable": true
                    },
                    "areas_experiencia": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Finanzas",
                            "Auditoría"
                        ],
                        "nullable": true
                    },
                    "competencias": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Contabilidad",
                            "Excel avanzado",
                            "SAP"
                        ],
                        "nullable": true
                    },
                    "rango_renta": {
                        "type": "string",
                        "example": "600k-900k",
                        "nullable": true
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "enum": [
                            "completa",
                            "part-time",
                            "por-horas"
                        ],
                        "example": "completa",
                        "nullable": true
                    },
                    "modalidad": {
                        "type": "string",
                        "enum": [
                            "presencial",
                            "remoto",
                            "hibrido"
                        ],
                        "example": "presencial",
                        "nullable": true
                    },
                    "cursos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "nombre": {
                                    "type": "string",
                                    "example": "Excel Avanzado"
                                },
                                "institucion": {
                                    "type": "string",
                                    "example": "SENCE"
                                },
                                "anio": {
                                    "type": "integer",
                                    "example": 2022
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "idiomas": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "idioma": {
                                    "type": "string",
                                    "example": "Inglés"
                                },
                                "nivel": {
                                    "type": "string",
                                    "enum": [
                                        "basico",
                                        "intermedio",
                                        "avanzado",
                                        "nativo"
                                    ],
                                    "example": "intermedio"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "portafolio_url": {
                        "type": "string",
                        "example": "https://miportafolio.cl",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "email": "maria@example.com",
                    "telefono": "+56912345678",
                    "resumen": "Contadora con experiencia en pymes y sector público",
                    "nivel_educacional": "universitaria",
                    "titulo_carrera": "Contador Auditor",
                    "anio_egreso": 2018,
                    "anios_experiencia": 7,
                    "rango_renta": "600k-900k",
                    "tipo_jornada": "completa",
                    "modalidad": "presencial",
                    "competencias": [
                        "Contabilidad",
                        "Excel avanzado",
                        "SAP"
                    ],
                    "areas_experiencia": [
                        "Finanzas",
                        "Auditoría"
                    ],
                    "persona_discapacidad": false
                }
            },
            "Persona": {
                "title": "Persona",
                "description": "Modelo completo de persona/talento",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "PROV-2026-A1B2"
                    },
                    "resumen": {
                        "type": "string",
                        "example": "Desarrollador web con 5 años de experiencia",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "enum": [
                            "basica",
                            "media",
                            "tecnica",
                            "universitaria",
                            "postgrado"
                        ],
                        "example": "universitaria",
                        "nullable": true
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "example": "Ingeniero en Informática",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "example": 2019,
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 5
                    },
                    "rango_renta": {
                        "type": "string",
                        "example": "800k-1.2M",
                        "nullable": true
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "enum": [
                            "completa",
                            "part-time",
                            "por-horas"
                        ],
                        "example": "completa",
                        "nullable": true
                    },
                    "modalidad": {
                        "type": "string",
                        "enum": [
                            "presencial",
                            "remoto",
                            "hibrido"
                        ],
                        "example": "hibrido",
                        "nullable": true
                    },
                    "portafolio_url": {
                        "type": "string",
                        "example": "https://miportafolio.cl",
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false
                    },
                    "validado": {
                        "type": "boolean",
                        "example": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "porcentaje_completitud": {
                        "type": "integer",
                        "example": 90
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-15T10:30:00+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-20T08:15:00+00:00"
                    }
                },
                "type": "object",
                "example": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "email": "juan@example.com",
                    "codigo_talento": "PROV-2026-A1B2",
                    "resumen": "Desarrollador web con 5 años de experiencia",
                    "nivel_educacional": "universitaria",
                    "titulo_carrera": "Ingeniero en Informática",
                    "anio_egreso": 2019,
                    "anios_experiencia": 5,
                    "rango_renta": "800k-1.2M",
                    "tipo_jornada": "completa",
                    "modalidad": "hibrido",
                    "persona_discapacidad": false,
                    "validado": true,
                    "activo": true,
                    "porcentaje_completitud": 90
                }
            }
        }
    },
...
}